CryptocurrencyPaymentAPI - Convertion of currencies using BitPay service

+++ CryptocurrencyPaymentAPI/DTOs/CurrencyRateDto.cs
@@ -3,7 +3,7 @@
     public class CurrencyRateDto
     {
         public string Currency { get; set; }
-        public float Amount { get; set; }
-        public float Rate { get; set; }
+        public double Amount { get; set; }
+        public double Rate { get; set; }
     }
 }
